requirements.txt
Flask
flask-cors
google-generativeai
PyMuPDF
python-dotenv